Isolated Matrix handling implementation
