+++ Stack/Opc.Ua.Core/Types/BuiltIn/Variant.cs
@@ -2604,248 +2604,4 @@ namespace Opc.Ua
     }//class
     #endregion
 
-    /// <summary>
-    /// Wraps a multi-dimensional array for use within a Variant.
-    /// </summary>
-    [DataContract(Namespace = Namespaces.OpcUaXsd)]
-    public class Matrix : IFormattable
-    {
-        #region Constructors
-        /// <summary>
-        /// Initializes the matrix with a multidimensional array.
-        /// </summary>
-        public Matrix(Array value, BuiltInType builtInType)
-        {
-            if (value == null) throw new ArgumentNullException(nameof(value));
-
-            m_elements = value;
-            m_dimensions = new int[value.Rank];
-
-            for (int ii = 0; ii < m_dimensions.Length; ii++)
-            {
-                m_dimensions[ii] = value.GetLength(ii);
-            }
-
-            m_elements = Utils.FlattenArray(value);
-            m_typeInfo = new TypeInfo(builtInType, m_dimensions.Length);
-
-#if DEBUG
-            TypeInfo sanityCheck = TypeInfo.Construct(m_elements);
-            System.Diagnostics.Debug.Assert(sanityCheck.BuiltInType == builtInType || (sanityCheck.BuiltInType == BuiltInType.ByteString && builtInType == BuiltInType.Byte));
-#endif
-        }
-
-        /// <summary>
-        /// Initializes the matrix with a one dimensional array and a list of dimensions.
-        /// </summary>
-        public Matrix(Array elements, BuiltInType builtInType, params int[] dimensions)
-        {
-            if (elements == null) throw new ArgumentNullException(nameof(elements));
-
-            m_elements = elements;
-            m_dimensions = dimensions;
-
-            if (dimensions != null && dimensions.Length > 0)
-            {
-                int length = 1;
-
-                try
-                {
-                    for (int ii = 0; ii < dimensions.Length; ii++)
-                    {
-                        checked
-                        {
-                            length *= dimensions[ii];
-                        }
-                    }
-                }
-                catch (OverflowException)
-                {
-                    throw new ArgumentException("The dimensions of the matrix are invalid and overflow when used to calculate the size.");
-                }
-
-                if (length != elements.Length)
-                {
-                    throw new ArgumentException("The number of elements in the array does not match the dimensions.");
-                }
-            }
-            else
-            {
-                m_dimensions = new int[] { elements.Length };
-            }
-
-            m_typeInfo = new TypeInfo(builtInType, m_dimensions.Length);
-
-#if DEBUG
-            TypeInfo sanityCheck = TypeInfo.Construct(m_elements);
-            System.Diagnostics.Debug.Assert(sanityCheck.BuiltInType == builtInType ||
-                (sanityCheck.BuiltInType == BuiltInType.Int32 && builtInType == BuiltInType.Enumeration) ||
-                (sanityCheck.BuiltInType == BuiltInType.ByteString && builtInType == BuiltInType.Byte) ||
-                (builtInType == BuiltInType.Variant));
-#endif
-        }
-        #endregion
-
-        #region Public Members
-        /// <summary>
-        /// The elements of the matrix.
-        /// </summary>
-        /// <value>An array of elements.</value>
-        public Array Elements => m_elements;
-
-        /// <summary>
-        /// The dimensions of the matrix.
-        /// </summary>
-        /// <value>The dimensions of the array.</value>
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Performance", "CA1819:PropertiesShouldNotReturnArrays")]
-        public int[] Dimensions => m_dimensions;
-
-        /// <summary>
-        /// The type information for the matrix.
-        /// </summary>
-        /// <value>The type information.</value>
-        public TypeInfo TypeInfo => m_typeInfo;
-
-        /// <summary>
-        /// Returns the flattened array as a multi-dimensional array.
-        /// </summary>
-        public Array ToArray()
-        {
-            Array array = Array.CreateInstance(m_elements.GetType().GetElementType(), m_dimensions);
-
-            int[] indexes = new int[m_dimensions.Length];
-
-            for (int ii = 0; ii < m_elements.Length; ii++)
-            {
-                array.SetValue(m_elements.GetValue(ii), indexes);
-
-                for (int jj = indexes.Length - 1; jj >= 0; jj--)
-                {
-                    indexes[jj]++;
-
-                    if (indexes[jj] < m_dimensions[jj])
-                    {
-                        break;
-                    }
-
-                    indexes[jj] = 0;
-                }
-            }
-
-            return array;
-        }
-        #endregion
-
-        #region Overridden Methods
-        /// <summary>
-        /// Determines if the specified object is equal to the object.
-        /// </summary>
-        /// <remarks>
-        /// Determines if the specified object is equal to the object.
-        /// </remarks>
-        public override bool Equals(object obj)
-        {
-            if (Object.ReferenceEquals(this, obj))
-            {
-                return true;
-            }
-
-            Matrix matrix = obj as Matrix;
-
-            if (matrix != null)
-            {
-                if (!m_typeInfo.Equals(matrix.TypeInfo))
-                {
-                    return false;
-                }
-                if (!Utils.IsEqual(m_dimensions, matrix.Dimensions))
-                {
-                    return false;
-                }
-                return Utils.IsEqual(m_elements, matrix.Elements);
-            }
-
-            return false;
-        }
-
-        /// <summary>
-        /// Returns a unique hashcode for the object.
-        /// </summary>
-        public override int GetHashCode()
-        {
-            if (m_elements != null)
-            {
-                return m_elements.GetHashCode();
-            }
-            if (m_typeInfo != null)
-            {
-                return m_typeInfo.GetHashCode();
-            }
-            if (m_dimensions != null)
-            {
-                return m_dimensions.GetHashCode();
-            }
-            return base.GetHashCode();
-        }
-        #endregion
-
-        #region IFormattable Members
-        /// <summary>
-        /// Returns the string representation of the object.
-        /// </summary>
-        /// <param name="format">(Unused) Always pass a NULL value</param>
-        /// <param name="formatProvider">The format-provider to use. If unsure, pass an empty string or null</param>
-        /// <returns>
-        /// A <see cref="System.String"/> containing the value of the current instance in the specified format.
-        /// </returns>
-        /// <remarks>
-        /// Returns the string representation of the object.
-        /// </remarks>
-        /// <exception cref="FormatException">Thrown when the 'format' argument is NOT null.</exception>
-        public string ToString(string format, IFormatProvider formatProvider)
-        {
-            if (format == null)
-            {
-                StringBuilder buffer = new StringBuilder();
-
-                buffer.AppendFormat("{0}[", m_elements.GetType().GetElementType().Name);
-
-                for (int ii = 0; ii < m_dimensions.Length; ii++)
-                {
-                    if (ii > 0)
-                    {
-                        buffer.Append(',');
-                    }
-
-                    buffer.AppendFormat(formatProvider, "{0}", m_dimensions[ii]);
-                }
-
-                buffer.AppendFormat(formatProvider, "]");
-
-                return buffer.ToString();
-            }
-
-            throw new FormatException(Utils.Format("Invalid format string: '{0}'.", format));
-        }
-        #endregion
-
-        #region ICloneable Members
-        /// <summary>
-        /// Makes a deep copy of the object.
-        /// </summary>
-        /// <returns>
-        /// A new object that is a copy of this instance.
-        /// </returns>
-        public new object MemberwiseClone()
-        {
-            return new Matrix((Array)Utils.Clone(m_elements), m_typeInfo.BuiltInType, (int[])Utils.Clone(m_dimensions));
-        }
-        #endregion
-
-        #region Private Fields
-        private Array m_elements;
-        private int[] m_dimensions;
-        private TypeInfo m_typeInfo;
-        #endregion
-    }
 }//namespace
