+++ Stack/Opc.Ua.Core/Types/Encoders/BinaryDecoder.cs
@@ -1508,49 +1508,9 @@ namespace Opc.Ua
                 Int32Collection dimensions = ReadInt32Array(null);
                 if (dimensions != null && dimensions.Count > 0)
                 {
-                    int length = 1;
+                    //int length;
+                    (_, int length) = Matrix.ValidateDimensions(dimensions, Context.MaxArrayLength);
 
-                    try
-                    {
-                        for (int ii = 0; ii < dimensions.Count; ii++)
-                        {
-                            if (dimensions[ii] <= 0)
-                            {
-                                /* The number of values is 0 if one or more dimension is less than or equal to 0.*/
-                                Utils.LogTrace("ReadArray read dimensions[{0}] = {1}. Matrix will have 0 elements.", ii, dimensions[ii]);
-                                dimensions[ii] = 0;
-                                length = 0;
-                                break;
-                            }
-                            else if (dimensions[ii] > m_context.MaxArrayLength)
-                            {
-                                throw ServiceResultException.Create(
-                                    StatusCodes.BadEncodingLimitsExceeded,
-                                    "ArrayDimensions [{0}] = {1} is greater than MaxArrayLength {2}.",
-                                    ii,
-                                    dimensions[ii],
-                                    m_context.MaxArrayLength);
-                            }
-
-                            checked
-                            {
-                                length *= dimensions[ii];
-                            }
-                        }
-                    }
-                    catch (OverflowException)
-                    {
-                        throw new ArgumentException("The dimensions of the matrix are invalid and overflow when used to calculate the size.");
-                    }
-                    if (length > m_context.MaxArrayLength)
-                    {
-                        throw ServiceResultException.Create(
-                            StatusCodes.BadEncodingLimitsExceeded,
-                            "Maximum array length of {0} was exceeded while summing up to {1} from the array dimensions",
-                            m_context.MaxArrayLength,
-                            length
-                            );
-                    }
                     // read the elements
                     Array elements = null;
                     if (encodeableTypeId != null)
@@ -2211,30 +2171,9 @@ namespace Opc.Ua
                         }
 
                         int[] dimensionsArray = dimensions.ToArray();
-                        int matrixLength = 1;
+                        (bool valid, int matrixLength) = Matrix.ValidateDimensions(dimensionsArray, length, Context.MaxArrayLength);
 
-                        for (int ii = 0; ii < dimensionsArray.Length; ii++)
-                        {
-                            if (dimensionsArray[ii] == 0 && length > 0)
-                            {
-                                throw new ServiceResultException(
-                                    StatusCodes.BadDecodingError,
-                                    Utils.Format("ArrayDimensions [{0}] is zero in Variant object.", ii));
-                            }
-                            else if (dimensionsArray[ii] > length && length > 0)
-                            {
-                                throw new ServiceResultException(
-                                    StatusCodes.BadDecodingError,
-                                    Utils.Format("ArrayDimensions [{0}] = {1} is greater than length {2}.", ii, dimensionsArray[ii], length));
-                            }
-
-                            checked
-                            {
-                                matrixLength *= dimensionsArray[ii];
-                            }
-                        }
-
-                        if (matrixLength != length)
+                        if (!valid || (matrixLength != length))
                         {
                             throw new ServiceResultException(StatusCodes.BadDecodingError, "ArrayDimensions does not match with the ArrayLength in Variant object.");
                         }
